requirements.txt
GitPython
requests
prompt-toolkit
termcolor
tabulate
pyfiglet